fix: check undefined meta var in fix
